wip
test/image/build-app.sh && qemu-system-x86_64 -bios /usr/share/edk2.git/ovmf-ia32/OVMF-pure-efi.fd -drive id=disk,file=try.img,if=none,format=raw -nic none -device ahci,id=ahci -device ide-hd,drive=disk,bus=ahci.0 -display none -serial mon:stdio -m 256

or drop -display none -serial mon:stdio

For Chrome OS:

test/image/build-app2.sh && qemu-system-x86_64 -bios /usr/share/edk2.git/ovmf-ia32/OVMF-pure-efi.fd -drive id=disk,file=volteer_test_image.bin,if=none,format=raw -nic none -device ahci,id=ahci -device ide-hd,drive=disk,bus=ahci.0  -m 256

crosfw -V efi-x86_app32

+++ lib/efi/efi.c
@@ -169,7 +169,8 @@ int efi_store_memory_map(struct efi_priv *priv)
 		puts(" Can't get memory map\n");
 		return ret;
 	}
-	printf("key=%x, image=%p\n", priv->memmap_key, priv->parent_image);
+	printf("key=%x, image=%p\n", (uint)priv->memmap_key,
+	       priv->parent_image);
 
 	return 0;
 }
